# docs/openapi.yaml
openapi: 3.0.3
info:
  title: E-Auction Platform API
  description: A microservices-based e-auction platform with polyglot persistence
  version: 1.0.0
  contact:
    name: E-Auction Support

servers:
  - url: http://localhost:8000/api
    description: Kong API Gateway (Development)

tags:
  - name: Auth
    description: Authentication and session management
  - name: Users
    description: User management
  - name: Products
    description: Product and auction management
  - name: Bids
    description: Bidding operations
  - name: Recommendations
    description: Product recommendations

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            statusCode:
              type: integer
            message:
              type: string
            error:
              type: string

    RegisterRequest:
      type: object
      required: [email, password, role, name]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 6
          example: password123
        role:
          type: string
          enum: [buyer, seller]
          example: buyer
        name:
          type: string
          example: John Doe

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: password123

    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: string
                email:
                  type: string
                role:
                  type: string
            tokens:
              type: object
              properties:
                accessToken:
                  type: string
                refreshToken:
                  type: string
                expiresIn:
                  type: integer

    RefreshTokenRequest:
      type: object
      required: [userId, refreshToken]
      properties:
        userId:
          type: string
        refreshToken:
          type: string

    Address:
      type: object
      required: [street, city, province, postalCode]
      properties:
        street:
          type: string
          example: 123 Main St
        city:
          type: string
          example: Jakarta
        province:
          type: string
          example: DKI Jakarta
        postalCode:
          type: string
          example: "12345"

    UserProfile:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: John Doe
        phone:
          type: string
          example: "+1234567890"
        address:
          $ref: '#/components/schemas/Address'
        avatar:
          type: string

    CreateUserRequest:
      type: object
      required: [email, password, role, profile]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          example: password123
        role:
          type: string
          enum: [buyer, seller]
        profile:
          $ref: '#/components/schemas/UserProfile'

    User:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
        role:
          type: string
        profile:
          $ref: '#/components/schemas/UserProfile'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateProductRequest:
      type: object
      required: [title, description, category, startingPrice, condition]
      properties:
        title:
          type: string
          example: Vintage Watch
        description:
          type: string
          example: A beautiful vintage watch
        category:
          type: string
          example: watches
        images:
          type: array
          items:
            type: string
        startingPrice:
          type: number
          example: 100
        buyNowPrice:
          type: number
          example: 500
        condition:
          type: string
          enum: [new, used]
        specifications:
          type: object
          additionalProperties: true

    Product:
      type: object
      properties:
        _id:
          type: string
        sellerId:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        images:
          type: array
          items:
            type: string
        startingPrice:
          type: number
        currentPrice:
          type: number
        buyNowPrice:
          type: number
        condition:
          type: string
        specifications:
          type: object
        auction:
          type: object
          properties:
            status:
              type: string
              enum: [draft, active, ended, cancelled]
            startTime:
              type: string
              format: date-time
            endTime:
              type: string
              format: date-time
            winnerId:
              type: string
            totalBids:
              type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    StartAuctionRequest:
      type: object
      required: [startTime, endTime]
      properties:
        startTime:
          type: string
          format: date-time
          example: "2025-12-03T00:00:00Z"
        endTime:
          type: string
          format: date-time
          example: "2025-12-10T00:00:00Z"

    CreateBidRequest:
      type: object
      required: [productId, amount]
      properties:
        productId:
          type: string
        amount:
          type: number
          example: 150

    Bid:
      type: object
      properties:
        bidId:
          type: string
        productId:
          type: string
        userId:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [active, outbid, winner]
        timestamp:
          type: integer

paths:
  /auth/register:
    post:
      operationId: registerUser
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email already exists

  /auth/login:
    post:
      operationId: loginUser
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: logoutUser
      tags: [Auth]
      summary: Logout user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully

  /auth/refresh-token:
    post:
      operationId: refreshToken
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /auth/verify:
    get:
      operationId: verifySession
      tags: [Auth]
      summary: Verify session
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Session valid
        '401':
          description: Invalid or expired session

  /users/{id}:
    get:
      operationId: getUserById
      tags: [Users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    put:
      operationId: updateUser
      tags: [Users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                profile:
                  $ref: '#/components/schemas/UserProfile'
      responses:
        '200':
          description: User updated
    delete:
      operationId: deleteUser
      tags: [Users]
      summary: Delete user (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted

  /products:
    get:
      operationId: listProducts
      tags: [Products]
      summary: List products with pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, ended, cancelled]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Products list
    post:
      operationId: createProduct
      tags: [Products]
      summary: Create a new product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

  /products/{id}:
    get:
      operationId: getProductById
      tags: [Products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    put:
      operationId: updateProduct
      tags: [Products]
      summary: Update product (draft only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Product updated
    delete:
      operationId: deleteProduct
      tags: [Products]
      summary: Delete product (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product deleted

  /products/seller/{sellerId}:
    get:
      operationId: getProductsBySeller
      tags: [Products]
      summary: Get products by seller
      parameters:
        - name: sellerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Products list

  /products/category/{category}:
    get:
      operationId: getProductsByCategory
      tags: [Products]
      summary: Get products by category
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Products list

  /products/{id}/auction-status:
    get:
      operationId: getAuctionStatus
      tags: [Products]
      summary: Get auction status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Auction status

  /products/{id}/start-auction:
    put:
      operationId: startAuction
      tags: [Products]
      summary: Start auction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAuctionRequest'
      responses:
        '200':
          description: Auction started

  /products/{id}/end-auction:
    put:
      operationId: endAuction
      tags: [Products]
      summary: End auction
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Auction ended

  /bids:
    post:
      operationId: placeBid
      tags: [Bids]
      summary: Place a bid
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBidRequest'
      responses:
        '201':
          description: Bid placed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bid'
        '400':
          description: Invalid bid amount
        '403':
          description: Seller cannot bid

  /bids/product/{productId}:
    get:
      operationId: getProductBids
      tags: [Bids]
      summary: Get bids for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bids list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Bid'

  /bids/product/{productId}/highest:
    get:
      operationId: getHighestBid
      tags: [Bids]
      summary: Get highest bid for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Highest bid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bid'

  /bids/product/{productId}/winner:
    get:
      operationId: getAuctionWinner
      tags: [Bids]
      summary: Get auction winner
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Winner bid
        '400':
          description: Auction not ended

  /bids/user/{userId}:
    get:
      operationId: getUserBids
      tags: [Bids]
      summary: Get user's bids
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User's bids

  /recommendations/user/{userId}:
    get:
      operationId: getUserRecommendations
      tags: [Recommendations]
      summary: Get recommendations for user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recommendations

  /recommendations/product/{productId}/similar:
    get:
      operationId: getSimilarProducts
      tags: [Recommendations]
      summary: Get similar products
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Similar products

  /recommendations/train:
    post:
      operationId: trainRecommendationModel
      tags: [Recommendations]
      summary: Train recommendation model
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Training started